Updated for week 2
Restructured file system to allow for multiple versions of the app in
the GH-Pages creation.

Created and updated additional pages as required by assignment in the
Gold app, including: About, Newsfeed, Search.

As part of the above, redid search bar.

Still having an issue with form, will correct in future update.
Will work on style difference for Bronze App.